Move GoogleAdsAccountCard to shared component folder
