+++ js/src/components/google-ads-account-card/authorize-ads.js
