+++ js/src/components/google-ads-account-card/connect-ads/index.scss
