+++ js/src/components/google-ads-account-card/terms-modal/index.scss
